code-style improvements and pylint fixes
